+++ examples/embedded-product/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @opengeni/example-embedded-product
 
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies [3d9ab25]
+  - @opengeni/react@5.0.4
+
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies [6a60a58]
+  - @opengeni/contracts@3.0.1
+  - @opengeni/react@5.0.3
+  - @opengeni/sdk@5.0.3
+
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies [05957ee]
+  - @opengeni/react@5.0.2
+
 ## 0.0.2
 
 ### Patch Changes