+++ packages/browserd/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @opengeni/browserd
 
+## 0.1.40
+
+### Patch Changes
+
+- Updated dependencies [6a60a58]
+  - @opengeni/contracts@3.0.1
+  - @opengeni/interaction@0.4.32
+
 ## 0.1.39
 
 ### Patch Changes