+++ packages/documents/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @opengeni/documents
 
+## 0.8.26
+
+### Patch Changes
+
+- Updated dependencies [a9cc903]
+  - @opengeni/db@4.3.2
+
+## 0.8.25
+
+### Patch Changes
+
+- Updated dependencies [6a60a58]
+  - @opengeni/db@4.3.1
+  - @opengeni/contracts@3.0.1
+  - @opengeni/config@1.1.1
+  - @opengeni/storage@0.2.126
+
 ## 0.8.24
 
 ### Patch Changes