+++ packages/sdk/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @opengeni/sdk
 
+## 5.0.3
+
+### Patch Changes
+
+- Updated dependencies [6a60a58]
+  - @opengeni/contracts@3.0.1
+
 ## 5.0.0
 
 ### Major Changes